requirements.txt
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
idna==2.10
mysql-connector-python==8.0.23
protobuf==3.15.5
requests==2.25.1
six==1.15.0
urllib3==1.26.3
